Update new runner acceptance test with reference payment feature
Now that we have the features confirmation email, reference number and payment link,
we need to ensure the Runner presents these features correctly.

The `new-runner-acceptance-test` form has been adjusted to enable confirmation email,
reference number and payment link. This commit adjusts the `NewRunner` spec, to
reflect this adjustment.

We check the confirmation page styling has changed, we also capture the
`reference_number` on this page to be used later to check the correct email (with pdf
and csv) attachments were sent on form submission.

Changes to the `find_pdf_attachments` method were made to allow us to pass in the
argument `expected_emails`, as we should expect 2 in the inbox if Confirmation
email is enabled, but only one if it is not (as is the case for the
`new_runner_branching` spec.