Provide an answer that passes email validation
We have changed the email address question from type `text` to type
`email`  to enable the confirmation email feature.

So we will now need to provide a valid email address for this page
to ensure we do not trigger the validation and allow us to move onto
the next question.