Run integration tests split into 4 parallel CI jobs
Tentative approach to running the integration tests (acceptance tests) split into 4 groups, each group running some logical tests (components, features v1, features v2 and lib) and each group running in its separate CI job, in parallel.

The assumption is, this may not save us time, but may reduce the amount of times we need to retry the whole run of the suite, as if 1 of the groups fail, we only need need to retry that failed job, instead of the whole suite.

This is still a WIP, started with this repo as it is manageable, but if it works we would need to reorg the job steps in the other repos too.