Making the finder by subject more specific
The submission emails will both contain "Submission from" in their subject, so we need to be more specific otherwise there is no certainty we will not be getting the CSV one, which will not contain the answers in the body.

In addition to this we can now also test the CSV email body to ensure it does not contain the answers or anything else (the body should be empty).