Cleanup legacy/unused save&return code
The tests were being excluded from the suite anyways:

`--exclude-pattern 'spec/features/v1/save_and_return_module_spec.rb'`

This is a cleanup of the tests and any related support code, including a simulator app, env variables, etc.

Not really related to this, but I've also added missing port mapping `18012:18012` to the `docker-compose.yml` because it is referenced in the wait_for_services script and in the Procfile.local

In theory nothing should be broken by this cleanup.