Bump ruby version and replace deprecated gem
Bump to latest ruby stable version 3.2.2
Closes #376

This is under the assumption the acceptance tests run in isolation so other projects can stay behind in an older ruby version for the time being, not forcing us to also update those.

Replaced deprecated `google-api-client` gem with `google-apis-gmail_v1` which is the only google service we need. Interface and code is the same so no changes to code are needed.
More info here:
https://googleapis.dev/ruby/google-api-client/latest/index.html

Decreased the timeouts to something more sensible now that we've implemented some retries in code. It means it fails quicker and retry instead of hanging for 180 secs.
However these time values need to be observed in case it increases test failures.