Legacy cleanup - part 1
https://trello.com/c/OGvzdwUH

First pass on legacy cleanup. More to come. Trying to remove as much as possible unused or obsolete code that was only needed for legacy forms.

Note the legacy test forms are still deployed in `formbuilder-services-test-dev`, and will be deleted (their deployments) once we don't use them in these acceptance tests.